wip: move unit tests out of src dir and update filters
