+++ packages/vite-plugin-svelte/__tests__/fixtures/preprocess/foo.scss
