+++ packages/vite-plugin-svelte/__tests__/svelte-version.spec.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import { compareToSvelte, atLeastSvelte, parseVersion } from '../svelte-version.js';
+import { compareToSvelte, atLeastSvelte, parseVersion } from '../src/utils/svelte-version.js';
 import { VERSION } from 'svelte/compiler';
 const svelteVersion = parseVersion(VERSION);
 describe('svelte-version', () => {
